Divide `CoursesScreen` into sub-components
