+++ src/components/Screen/CoursesScreen/index.tsx
@@ -1,33 +1,25 @@
-import React, { ReactElement, useState, useEffect, Key } from 'react';
-import { Divider, message, ConfigProvider } from 'antd';
+import React, { ReactElement, useEffect, Key } from 'react';
+import { Divider, ConfigProvider } from 'antd';
 
 import { coursesMockData } from '../../../constants/mocks/courses-screen';
-import {
-  useCoursesContext,
-  fetchCoursesData,
-  ActionType,
-  exportCourses,
-  deleteCourses,
-} from '../../../providers/CoursesContext';
-import { UICourse, PublishStatus } from '../../../walkme/data';
+import { useCoursesContext, fetchCoursesData, ActionType } from '../../../providers/CoursesContext';
+import { UICourse } from '../../../walkme/data';
 import { IDateRange } from '../../../utils';
 
 import AnalyticsCharts from '../../common/AnalyticsCharts';
 import ControlsWrapper from '../../common/ControlsWrapper';
-import { ExportButton, CreateButton } from '../../common/buttons';
+import { CreateButton } from '../../common/buttons';
 import Icon, { IconType } from '../../common/Icon';
 import ScreenHeader from '../../common/ScreenHeader';
-import { SearchFilter } from '../../common/filters';
-import WMButton, { ButtonVariantEnum } from '../../common/WMButton';
+// import { DropdownFilter } from '../../common/filters';
 import WMCard from '../../common/WMCard';
 import WMTable from '../../common/WMTable';
-import {
-  DeleteCourseDialog,
-  CantDeleteDialog, // TODO: remove 'delete' button from this dialog, disabled for now
-  PublishToEnvironmentDialog,
-  ExportToCSVDialog,
-} from '../../common/dialogs';
 
+import ShownCoursesIndicator from './ShownCoursesIndicator';
+import ProductionStatusActions from './ProductionStatusActions';
+import DeleteCoursesButton from './DeleteCoursesButton';
+import ExportCoursesButton from './ExportCoursesButton';
+import SearchCoursesFilter from './SearchCoursesFilter';
 // import { statuses, segments } from './utils';
 import { columns } from './tableData';
 import classes from './style.module.scss';
@@ -38,7 +30,6 @@ export default function CoursesScreen(): ReactElement {
   const [state, dispatch] = useCoursesContext();
   const {
     dateRange: { from, to },
-    courses,
     overview,
     filteredCourses,
     selectedRows,
@@ -51,32 +42,6 @@ export default function CoursesScreen(): ReactElement {
 
   // Unmount only
   useEffect(() => () => dispatch({ type: ActionType.ResetCourses }), [dispatch]);
-
-  const onSearch = (searchValue: string) => {
-    const newCourseList = courses.filter(({ title }) =>
-      title.toLowerCase().includes(searchValue.toLowerCase()),
-    );
-
-    dispatch({
-      type: ActionType.SetCoursesSearchValue,
-      coursesSearchValue: searchValue,
-      courses: newCourseList,
-    });
-  };
-
-  // Dialogs
-  const [showPublish, setShowPublish] = useState(false);
-  const [showExport, setShowExport] = useState(false);
-  const [showDeleteCourse, setShowDeleteCourse] = useState(false);
-  const [showCantDeleteCourse, setShowCantDeleteCourse] = useState(false);
-
-  const checkCantDelete = () =>
-    selectedRows.some((course) => course.publishStatus === PublishStatus.Published);
-
-  const onDeleteCourse = () => {
-    if (checkCantDelete()) setShowCantDeleteCourse(true);
-    else setShowDeleteCourse(true);
-  };
 
   const onMultiSelectChange = (selectedRowKeys: Array<Key>, selectedRows: Array<UICourse>) =>
     dispatch({
@@ -98,7 +63,6 @@ export default function CoursesScreen(): ReactElement {
     dispatch({ type: ActionType.SetDateRange, dateRange });
 
   const selectedRowsCount = selectedRows.length;
-  const shownCoursesCount = filteredCourses.length;
 
   return (
     <>
@@ -121,15 +85,7 @@ export default function CoursesScreen(): ReactElement {
             columns={columns}
             // onSortEnd={(sortedData) => setTableData(sortedData)}
           >
-            <div className={classes['showing']}>
-              {courses.length ? (
-                <>
-                  {selectedRowsCount
-                    ? `${selectedRowsCount} course${selectedRowsCount > 1 ? 's' : ''} selected`
-                    : `Showing ${shownCoursesCount} course${shownCoursesCount > 1 ? 's' : ''}`}
-                </>
-              ) : null}
-            </div>
+            <ShownCoursesIndicator />
             {/* <ControlsWrapper>
               <DropdownFilter label="Status" options={statuses} />
               <DropdownFilter label="Segments" options={segments} />
@@ -137,75 +93,18 @@ export default function CoursesScreen(): ReactElement {
             <ControlsWrapper>
               {selectedRowsCount ? (
                 <>
-                  <div className={classes['prod-status-btns']}>
-                    <WMButton variant={ButtonVariantEnum.Link} onClick={() => setShowPublish(true)}>
-                      Publish
-                    </WMButton>
-                    <WMButton
-                      variant={ButtonVariantEnum.Link}
-                      onClick={() => message.info(`Production status was changed to archive`)}
-                    >
-                      Archive
-                    </WMButton>
-                    <WMButton
-                      variant={ButtonVariantEnum.Link}
-                      onClick={() => message.info(`Production status was changed to draft`)}
-                    >
-                      Mark as Draft
-                    </WMButton>
-                  </div>
-                  <WMButton
-                    className={classes['delete-btn']}
-                    icon={<Icon type={IconType.Delete} />}
-                    onClick={onDeleteCourse}
-                  />
+                  <ProductionStatusActions />
+                  <DeleteCoursesButton />
                   <Divider className={classes['separator']} type="vertical" />
                 </>
               ) : null}
-              <ExportButton onClick={() => setShowExport(true)} />
-              <SearchFilter
-                className={classes['search']}
-                placeholder="Search course name"
-                onSearch={onSearch}
-              />
+              <ExportCoursesButton />
+              <SearchCoursesFilter />
               <CreateButton />
             </ControlsWrapper>
           </WMTable>
         </ConfigProvider>
       </WMCard>
-      {/* Dialogs */}
-      <PublishToEnvironmentDialog
-        open={showPublish}
-        onCancel={() => setShowPublish(false)}
-        onConfirm={() => {
-          setShowPublish(false);
-          message.info(`Production status changed to Published`);
-        }}
-      />
-      <DeleteCourseDialog
-        courses={selectedRows}
-        open={showDeleteCourse}
-        onCancel={() => setShowDeleteCourse(false)}
-        onConfirm={async () => {
-          setShowDeleteCourse(false);
-          await deleteCourses(dispatch, selectedRows);
-          fetchCoursesData(dispatch, 0, from, to);
-        }}
-      />
-      <CantDeleteDialog
-        open={showCantDeleteCourse}
-        onCancel={() => setShowCantDeleteCourse(false)}
-        onConfirm={() => setShowCantDeleteCourse(false)}
-      />
-      <ExportToCSVDialog
-        coursesCount={courses.length}
-        open={showExport}
-        onCancel={() => setShowExport(false)}
-        onConfirm={() => {
-          setShowExport(false);
-          exportCourses(dispatch, 0, from, to);
-        }}
-      />
     </>
   );
 }
